+++ src/main/kotlin/dev/minn/jda/ktx/messages/builder.kt
@@ -14,14 +14,9 @@
  * limitations under the License.
  */
 
-@file:Suppress("FunctionName")
-
-package dev.minn.jda.ktx
+package dev.minn.jda.ktx.messages
 
 import dev.minn.jda.ktx.interactions.row
-import dev.minn.jda.ktx.messages.Components
-import dev.minn.jda.ktx.messages.Embeds
-import dev.minn.jda.ktx.messages.allOf
 import net.dv8tion.jda.api.EmbedBuilder
 import net.dv8tion.jda.api.MessageBuilder
 import net.dv8tion.jda.api.entities.Message
